Move User model to baselayer; fix broken imports
